+++ src/Integration.php
@@ -1,5 +1,14 @@
 <?php
 
+/**
+ * Title: Mollie integration
+ * Description:
+ * Copyright: Copyright (c) 2005 - 2016
+ * Company: Pronamic
+ *
+ * @author Remco Tolsma
+ * @version 1.0.0
+ */
 class Pronamic_WP_Pay_Gateways_Mollie_Integration extends Pronamic_WP_Pay_Gateways_AbstractIntegration {
 	public function __construct() {
 		$this->id            = 'mollie';