+++ src/Listener.php
@@ -3,8 +3,9 @@
 /**
  * Title: Mollie listener
  * Description:
- * Copyright: Copyright (c) 2005 - 2015
+ * Copyright: Copyright (c) 2005 - 2016
  * Company: Pronamic
+ *
  * @author Remco Tolsma
  * @version 1.0.0
  */
@@ -19,7 +20,7 @@ class Pronamic_WP_Pay_Gateways_Mollie_Listener implements Pronamic_Pay_Gateways_
 
 			$payment = get_pronamic_payment_by_transaction_id( $transaction_id );
 
-			Pronamic_WP_Pay_Plugin::update_payment( $payment );
+			Pronamic_WP_Pay_Plugin::update_payment( $payment, false );
 		}
 	}
 }