+++ src/Settings.php
@@ -3,8 +3,9 @@
 /**
  * Title: Mollie gateway settings
  * Description:
- * Copyright: Copyright (c) 2005 - 2015
+ * Copyright: Copyright (c) 2005 - 2016
  * Company: Pronamic
+ *
  * @author Remco Tolsma
  * @version 1.2.0
  * @since 1.2.0